seq_1

+++ tasks/seq/kudryashova_i_vector_dot_product/src/vectorDotProductSeq.cpp
@@ -17,7 +17,7 @@ std::vector<int> kudryashova_i_vector_dot_product::GetRandomVector(const int siz
 
 int kudryashova_i_vector_dot_product::vectorDotProduct(const std::vector<int>& vector1, const std::vector<int>& vector2) {
     long long result = 0;
-    for (int i = 0; i < vector1.size(); ++i) {
+    for (unsigned long i = 0; i < vector1.size(); ++i) {
         result += vector1[i] * vector2[i];
     }
     return result;
@@ -27,7 +27,7 @@ bool kudryashova_i_vector_dot_product::TestTaskSequential::pre_processing() {
     internal_order_test();
 
     input_.resize(taskData->inputs.size());
-    for (int i = 0; i < input_.size(); ++i) {
+    for (unsigned long i = 0; i < input_.size(); ++i) {
         auto* tempPtr = reinterpret_cast<int*>(taskData->inputs[i]);
         input_[i] = std::vector<int>(taskData->inputs_count[i]);
         std::copy(tempPtr, tempPtr + taskData->inputs_count[i], input_[i].begin());
@@ -47,7 +47,7 @@ bool kudryashova_i_vector_dot_product::TestTaskSequential::validation() {
 
 bool kudryashova_i_vector_dot_product::TestTaskSequential::run() {
     internal_order_test();
-    for (int i = 0; i < input_[0].size(); i++) {
+    for (unsigned long i = 0; i < input_[0].size(); i++) {
         result += input_[1][i] * input_[0][i];
     }
     return true;
